factor out another test
